refactoring controller and services
