+++ dto/rest/response.py
@@ -1,0 +1,4 @@
+class Response:
+    def __init__(self, status, message):
+        self.status = status
+        self.message = message
